TYPE: Split old plain hints into "parameter" and "type"
`RsPlainParameterHint` contains only parameter name hints and enabled on
both 191 and 192.
`RsPlainTypeHint` contains only type hints and enabled only on 191
because of new interactive type hints on 192
